Moved use-export-to-gltf to hooks folder
